+++ crates/rpc/rpc-types-compat/Cargo.toml
@@ -12,5 +12,5 @@ repository.workspace = true
 [dependencies]
 reth-primitives.workspace = true
 reth-rpc-types.workspace = true
-
 serde = { workspace = true, features = ["derive"] }
+reth-rlp.workspace = true
