+++ crates/rpc/rpc-types-compat/src/lib.rs
@@ -1,1 +1,5 @@
-mod transaction;
+
+pub mod transaction;
+pub mod block;
+pub use transaction::*;
+pub use block::*;
