--- crates/rpc/rpc-types-compat/src/mod.rs
@@ -1,3 +1,0 @@
-pub mod transaction;
-
-pub use transaction::*;
