+++ crates/rpc/rpc-types/Cargo.toml
@@ -14,7 +14,6 @@ Reth RPC types
 # reth
 reth-primitives.workspace = true
 reth-rlp.workspace = true
-reth-rpc-types-compat.workspace = true
 # errors
 thiserror.workspace = true
