+++ crates/rpc/rpc/src/eth/api/block.rs
@@ -10,9 +10,9 @@ use crate::{
 use reth_network_api::NetworkInfo;
 use reth_primitives::{BlockId, BlockNumberOrTag, TransactionMeta};
 use reth_provider::{BlockReaderIdExt, EvmEnvProvider, StateProviderFactory};
-use reth_rpc_types::{Block, Index, RichBlock, TransactionReceipt};
+use reth_rpc_types::{ Index, RichBlock, TransactionReceipt};
 use reth_transaction_pool::TransactionPool;
-
+use reth_rpc_types_compat::block::{uncle_block_from_header,from_block};
 impl<Provider, Pool, Network> EthApi<Provider, Pool, Network>
 where
     Provider: BlockReaderIdExt + StateProviderFactory + EvmEnvProvider + 'static,
@@ -49,7 +49,7 @@ where
         let uncle = uncles
             .into_iter()
             .nth(index)
-            .map(|header| Block::uncle_block_from_header(header).into());
+            .map(|header| uncle_block_from_header(header).into());
         Ok(uncle)
     }
 
@@ -160,7 +160,7 @@ where
             .header_td_by_number(block.number)?
             .ok_or(EthApiError::UnknownBlockNumber)?;
         let block =
-            Block::from_block(block.into(), total_difficulty, full.into(), Some(block_hash))?;
+            from_block(block.into(), total_difficulty, full.into(), Some(block_hash))?;
         Ok(Some(block.into()))
     }
 }
